+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 8.6
 -----
-- Playlists: Add episode to a playlist from episode details [##3943](https://github.com/Automattic/pocket-casts-ios/pull/3943/) 
+- Playlists: Add episodes to a playlist from multi-select overflow menu [#4001](https://github.com/Automattic/pocket-casts-ios/pull/4002)
+- Playlists: Add episode to a playlist from episode details [##3943](https://github.com/Automattic/pocket-casts-ios/pull/3943/)
 - Improve cache invalidation for Playlist Count and Artwork [#3927](https://github.com/Automattic/pocket-casts-ios/pull/3927)
 - Dynamic Type: Update UpNext [#3918](https://github.com/Automattic/pocket-casts-ios/pull/3918)
 - Dynamic Type: Update Podcasts list [#3914](https://github.com/Automattic/pocket-casts-ios/pull/3914)
@@ -39,7 +40,9 @@
 - Dynamic type: Player effects [#3998](https://github.com/Automattic/pocket-casts-ios/pull/3998)
 - Dynamic type: Player sleep timer [#3997](https://github.com/Automattic/pocket-casts-ios/pull/3997)
 - Dynamic type: Discover Category [#3995](https://github.com/Automattic/pocket-casts-ios/pull/3995)
-
+- Dynamic type: Update multi select options menu [#4005](https://github.com/Automattic/pocket-casts-ios/pull/4005)
+- Dynamic type: Change email [#4009](https://github.com/Automattic/pocket-casts-ios/pull/4009)
+- Dynamic Type: Update Change password [#4011](https://github.com/Automattic/pocket-casts-ios/pull/4011)
 
 8.5
 -----